# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Production-ready backend service scaffolded by crank.",
        "title": "agni API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/apps": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists microVM applications belonging to the user or all cluster apps.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apps"
                ],
                "summary": "List applications",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_adapters_http_web_v1.appDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_surgged_agni_internal_adapters_http_web_rest.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new application with a presigned upload URL for archive submission.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apps"
                ],
                "summary": "Create application",
                "parameters": [
                    {
                        "description": "App creation request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http_web_v1.createAppDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http_web_v1.createAppResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_surgged_agni_internal_adapters_http_web_rest.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_surgged_agni_internal_adapters_http_web_rest.Error"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/github_com_surgged_agni_internal_adapters_http_web_rest.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/apps/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns details of a specific microVM application by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apps"
                ],
                "summary": "Get application details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http_web_v1.appDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_surgged_agni_internal_adapters_http_web_rest.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Destroys and cleans up a microVM application container and resources.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apps"
                ],
                "summary": "Destroy application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_surgged_agni_internal_adapters_http_web_rest.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/apps/{id}/deploy": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Starts the deployment workflow for an app that has an uploaded archive.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apps"
                ],
                "summary": "Deploy application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http_web_v1.deployResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_surgged_agni_internal_adapters_http_web_rest.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_surgged_agni_internal_adapters_http_web_rest.Error"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_surgged_agni_internal_adapters_http_web_rest.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/apps/{id}/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Streams live stdout/stderr logs from the microVM container via Server-Sent Events (SSE).",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "apps"
                ],
                "summary": "Stream application logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_adapters_http_web_v1.logLinePayload"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_surgged_agni_internal_adapters_http_web_rest.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/apps/{id}/multipart/abort": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancels an incomplete multipart upload and cleans up partial parts.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apps"
                ],
                "summary": "Abort multipart upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Upload ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "upload_id": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_surgged_agni_internal_adapters_http_web_rest.Error"
                        }
                    }
                }
            }
        },
        "/api/v1/apps/{id}/multipart/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Finalizes a multipart upload by assembling all uploaded parts.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apps"
                ],
                "summary": "Complete multipart upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Upload ID and completed parts",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "parts": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/github_com_surgged_agni_internal_ports.UploadedPart"
                                    }
                                },
                                "upload_id": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_surgged_agni_internal_adapters_http_web_rest.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_surgged_agni_internal_adapters_http_web_rest.Error": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "error": {
                    "type": "string",
                    "example": "validation failed"
                }
            }
        },
...
        "internal_adapters_http_web_v1.appDTO": {
            "type": "object",
            "required": [
                "name",
                "owner_email"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "failed_step": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_ref": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_email": {
                    "type": "string"
                },
                "pod_name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "runtime": {
                    "type": "string"
                },
                "service_url": {
                    "type": "string"
                },
                "share_url": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "internal_adapters_http_web_v1.deployResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\" to authenticate.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}